00_HACKATHON-SUBMISSIONS/0002_HealthyMe/server/requirements.txt
Flask
flask-cors
PyPDF2
google.generativeai
gunicorn
